split RecommendationsPage and LWCuratedPage
